+++ cli/integration-test/integration/stubs_test.clj
@@ -16,16 +16,10 @@
                                                  :stubs {:generation {:namespaces #{"datomic.api"}}})}))
   (lsp/notify! (fixture/initialized-notification))
   (Thread/sleep 20000) ;; Wait for async stubs generation
-  (lsp/notify! (fixture/did-open-source-path-notification "stubs/a.clj"))
 
-  (testing "After stub generation we find datomic.api analysis and diagnostics."
-    (h/assert-submaps
-      [{:range {:start {:line 5 :character 2} :end {:line 5 :character 21}}
-        :severity 1
-        :code "invalid-arity"
-        :source "clj-kondo"
-        :message "datomic.api/create-database is called with 0 args but expects 1"
-        :tags []}]
-      (lsp/client-awaits-server-diagnostics "stubs/a.clj"))
-
+  (testing "After stub generation we find datomic.api analysis but no diagnostics."
+    ;; TODO how to assert there is no diagnostics via protocol only?
+    ;; (h/assert-submaps
+    ;;   []
+    ;;   (lsp/client-awaits-server-diagnostics "stubs/a.clj")
     (h/delete-project-file "../../.lsp/.cache/stubs")))