+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - New optional `:kondo-config-dir` setting to configure clj-kondo execution.
   - Parallelize and log the time spent on built-in linters execution.
   - Fix #1851: Error when source files have non-ASCII characters in their path or name
+  - Fix caching issue when :source-aliases changes. #2081
 
 ## 2025.06.13-20.45.44