+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Parallelize and log the time spent on built-in linters execution.
   - Fix #1851: Error when source files have non-ASCII characters in their path or name
   - Fix caching issue when :source-aliases changes. #2081
+  - Fix emitting invalid messages if there's an internal error.
 
 ## 2025.06.13-20.45.44